L1_cloud: Updated I/O to read/write ground_pixel_quality_flag in netCDF as a 32-bit UINT, whereas OMI uses 16-bit. Also added a subroutine to read the observation month from the time_coverage_start attribute of netCDF files, and a test to flag pixels with extreme viewing zenith angles as bad.

+++ L1_cloud/src/m_vars.f90
@@ -93,7 +93,7 @@ module m_vars
 
   ! FLAGS
   logical, allocatable, dimension (:) :: land_flg ! pixel on land?
-  INTEGER (KIND=2), DIMENSION(:,:), allocatable :: geoflg ! geolocation flags
+  INTEGER (KIND=4), DIMENSION(:,:), allocatable :: geoflg ! geolocation flags
   INTEGER (KIND=1), DIMENSION(:,:), allocatable :: anomflg ! cross-track
   ! quality flags
   INTEGER (KIND=2), dimension(:), allocatable :: mflg ! measurement
